apps/phoenix-agent-ia/main.py
"""
🎯 Phoenix Agent IA - Clean Reset
Service d'agents IA spécialisés (consciouness, security, smart-router)
"""

import os
from datetime import datetime
from typing import Dict, Any, Optional, List
from fastapi import FastAPI, HTTPException, status, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel
from dotenv import load_dotenv

# Load environment
load_dotenv()

# Configuration
API_SECRET = os.getenv("API_SECRET_TOKEN", "dev-token")
ENVIRONMENT = os.getenv("ENVIRONMENT", "development")

# Security
security = HTTPBearer()

# FastAPI app
app = FastAPI(
    title="Phoenix Agent IA",
    description="Service d'agents IA spécialisés (Clean Reset)",
    version="1.0.0"
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"] if ENVIRONMENT == "development" else [],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class AgentRequest(BaseModel):
    agent_type: str  # consciousness, security, smart-router, data-flywheel
    task: str
    data: Optional[Dict[str, Any]] = None
    user_id: Optional[str] = None

class AgentResponse(BaseModel):
    success: bool
    agent_type: str
    result: Optional[Dict[str, Any]] = None
    message: str
    timestamp: datetime
    processing_time: Optional[float] = None

class HealthCheck(BaseModel):
    status: str
    timestamp: datetime
    active_agents: List[str]
    environment: str

# Auth dependency
async def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    if ENVIRONMENT == "development":
        return True  # Skip auth in dev
    
    if credentials.credentials != API_SECRET:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication token",
            headers={"WWW-Authenticate": "Bearer"},
        )
    return True

# Agent implementations (placeholders)
class ConsciousnessAgent:
    """Agent de conscience du système"""
    
    def process(self, task: str, data: Dict[str, Any]) -> Dict[str, Any]:
        # TODO: Implémenter logique de conscience système
        return {
            "consciousness_level": "aware",
            "system_health": "optimal",
            "recommendations": [
                "System is functioning normally",
                "All agents are responsive"
            ]
        }

class SecurityGuardian:
    """Agent de sécurité"""
    
    def process(self, task: str, data: Dict[str, Any]) -> Dict[str, Any]:
        # TODO: Implémenter validation sécurité
        return {
            "threat_level": "low",
            "security_score": 95,
            "alerts": [],
            "recommendations": [
                "No security threats detected",
                "All systems secure"
            ]
        }

class SmartRouter:
    """Agent de routage intelligent"""
    
    def process(self, task: str, data: Dict[str, Any]) -> Dict[str, Any]:
        # TODO: Implémenter routage intelligent
        return {
            "routing_decision": "backend-unified",
            "confidence": 0.95,
            "alternative_routes": ["iris-api"],
            "load_balancing": "optimal"
        }

class DataFlywheel:
    """Agent de traitement de données"""
    
    def process(self, task: str, data: Dict[str, Any]) -> Dict[str, Any]:
        # TODO: Implémenter traitement données
        return {
            "data_processed": True,
            "records_count": len(data.get("records", [])),
            "processing_status": "completed",
            "insights": ["Data pipeline running smoothly"]
        }

# Agent instances
agents = {
    "consciousness": ConsciousnessAgent(),
    "security": SecurityGuardian(),
    "smart-router": SmartRouter(),
    "data-flywheel": DataFlywheel()
}

# Routes
@app.get("/", response_model=Dict[str, str])
async def root():
    return {
        "service": "Phoenix Agent IA",
        "status": "operational",
        "version": "1.0.0 (clean reset)",
        "available_agents": list(agents.keys())
    }

@app.get("/health", response_model=HealthCheck)
async def health_check():
    return HealthCheck(
        status="healthy",
        timestamp=datetime.now(),
        active_agents=list(agents.keys()),
        environment=ENVIRONMENT
    )

@app.post("/agent/process", response_model=AgentResponse)
async def process_agent_task(
    request: AgentRequest,
    _: bool = Depends(verify_token)
):
    """Traite une tâche avec un agent spécialisé"""
    
    start_time = datetime.now()
    
    if request.agent_type not in agents:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=f"Unknown agent type: {request.agent_type}. Available: {list(agents.keys())}"
        )
    
    try:
        agent = agents[request.agent_type]
        result = agent.process(request.task, request.data or {})
        
        processing_time = (datetime.now() - start_time).total_seconds()
        
        return AgentResponse(
            success=True,
            agent_type=request.agent_type,
            result=result,
            message=f"Task processed successfully by {request.agent_type} agent",
            timestamp=datetime.now(),
            processing_time=processing_time
        )
        
    except Exception as e:
        processing_time = (datetime.now() - start_time).total_seconds()
        
        return AgentResponse(
            success=False,
            agent_type=request.agent_type,
            result=None,
            message=f"Error processing task: {str(e)}",
            timestamp=datetime.now(),
            processing_time=processing_time
        )

@app.get("/agents")
async def list_agents():
    """Liste tous les agents disponibles"""
    return {
        "agents": [
            {
                "name": "consciousness",
                "description": "Agent de conscience du système",
                "status": "active"
            },
            {
                "name": "security",
                "description": "Agent de sécurité et validation",
                "status": "active"
            },
            {
                "name": "smart-router",
                "description": "Agent de routage intelligent",
                "status": "active"
            },
            {
                "name": "data-flywheel",
                "description": "Agent de traitement de données",
                "status": "active"
            }
        ]
    }

@app.post("/consciousness/check")
async def consciousness_check(_: bool = Depends(verify_token)):
    """Endpoint spécifique pour check de conscience"""
    agent = agents["consciousness"]
    result = agent.process("system_check", {})
    
    return {
        "consciousness_status": "active",
        "system_awareness": result,
        "timestamp": datetime.now()
    }

@app.post("/security/scan")
async def security_scan(
    scan_data: Dict[str, Any],
    _: bool = Depends(verify_token)
):
    """Endpoint spécifique pour scan sécurité"""
    agent = agents["security"]
    result = agent.process("security_scan", scan_data)
    
    return {
        "scan_results": result,
        "timestamp": datetime.now()
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8002)