apps/phoenix-agent-ia/pyproject.toml
[tool.poetry]
name = "phoenix-agent-ia"
version = "1.0.0"
description = "Phoenix Agent IA - Clean Reset"
authors = ["Phoenix Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"